+++ lib/src/exceptions/module/module_exceptions.dart
@@ -1,6 +1,9 @@
-// Copyright (C) 2022-2023 Intel Corporation
+// Copyright (C) 2022-2024 Intel Corporation
 // SPDX-License-Identifier: BSD-3-Clause
 
+export 'invalid_hierarchy_exception.dart';
 export 'module_not_built_exception.dart';
 export 'port_does_not_exist_exception.dart';
+export 'port_rules_violation_exception.dart';
+export 'port_type_exception.dart';
 export 'port_width_mismatch_exception.dart';